+++ src/fft.h
@@ -1,7 +1,7 @@
 #ifndef __FFT_H
 #define __FFT_H
 
-void FFT(short int dir,long m,int16_t *x,int16_t *y); 
+void FFT(short int dir,long m,float *x,float *y); 
 
 
 #endif
