+++ src/lin_fit.h
@@ -10,10 +10,10 @@
 #define RP_BUF_SIZE 16384
 
 
-void MakeDomain(int bufsize, int16_t *ip);
+void MakeDomain(int bufsize, float *ip);
 
-int LinearFilter(int16_t * domain, int16_t * data, size_t n, int16_t c0, int16_t c1);
+int LinearFilter(float * domain, float * data, size_t n, float c0, float c1);
 
-int gsl_fit_linear (const int16_t *x, const int16_t *y, const size_t n, int16_t * c0, int16_t * c1);
+int gsl_fit_linear (const float *x, const float *y, const size_t n, float * c0, float * c1);
 
 #endif
