+++ src/window.h
@@ -2,8 +2,8 @@
 #define __WINDOW_H
 
 
-void GenWindow(int bufsize, int16_t *hp); /*generate hamming window vals */
+void GenWindow(int bufsize, float *hp); /*generate hamming window vals */
 
-void Hadamard(int bufsize, int16_t *hp, int16_t *dp); /* componentwise multiply the two arrays and store the result in hp*/ 
+void Hadamard(int bufsize, float *hp, float *dp); /* componentwise multiply the two arrays and store the result in hp*/ 
 
 #endif
