+++ src/write.h
@@ -3,11 +3,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int WriteTimeData(FILE * fd, double *time_series, int len);
+int WriteTimeData(FILE * fd, float *time_series, int len);
 
-int WriteTimeFreqData(FILE * fd, int16_t *time_series, int16_t * realf, int16_t * imf, int len);
+int WriteTimeFreqData(FILE * fd, float *time_series, float * realf, float * imf, int len);
 
-int WriteAllData(FILE *fd, int16_t * time_series, int16_t * realf, int16_t *imf, int16_t *cp, int16_t *icp); 
+int WriteAllData(FILE *fd, float * time_series, float * realf, float *imf, float *cp, float *icp); 
 
 
 #endif
